Fix files overwriting self during upload.
This was due to our OpenOptions in the disk provider - by default,
`.write(true)` will overwrite the existing file contents before
writing. This current fix is quick for my own testing purposes, but we
should check if the file already exists and act on that (maybe append a
hash to the end?).